Bizarre issues seem fixed on Android
It couldn't find any assets, despite them being inside the APK. MenuScreen code has moved from the constructor to the show() method because I thought maybe it was loading files in the constructor before it could use them. Android SDK version is now 28 because it's the most recent, and it should work on phones with as low as Android SDK 8.